location: fix TimezoneMonitor use-after-free
When a TimezoneMonitor is finalized, it triggers a GCancellable in
attempt to cancel ongoing asynchronous geoclue operations, right before
the TimezoneMonitor is freed.

However, cancellation here is asynchronous, it is not processed until
later. Two such callbacks here were attempting to access the
TimezoneMonitor in question before verifying that the geoclue operation
completed successfully (in event of finalization, the TimezoneMonitor is
already freed).

Do appropriate error checking in these callbacks before accessing
the TimezoneMonitor and its private data.

[endlessm/eos-shell#2560]