keyboard: disconnect from settings notification when finalizing
Due to the use of g_settings_delay() in the keyboard page, it's
possible for the backend object to survive the page itself, when the
locale changes.
In that scenario, a callback will be triggered with an invalid page
pointer, which will eventually lead to a crash.

Disconnect all handlers connected by the page when finalizing it.

[endlessm/eos-shell#2904]