keyboard: ignore locales without a language name
Adding to the internal hash table every language returned by
gnome_get_all_locales() without checking if they have a non-NULL visual
name to display will cause a crash when trying to use the search box,
since the call to strstr() in match_all, called from list_filter() will
segfault with the NULL string.

To avoid this issue, we just check early if there's an untranslated
name available for the locale in get_locale_infos(), not adding it to
the internal hash table and continuing to the next one if not the case.

[endlessm/eos-shell#3480]