Only show en_US and user languages (if any) in the initial page
This way, during the FBE the Global image should show only the en_US
language, while other images with personality would add whatever the
default language is defined in /etc/EndlessOS/personality.conf.

Note that more options will be shown if there are other users in the
system with aditional languages installed, but that should not affect
the FBE anyway.

[endlessm/eos-shell#3507]