Disable automatic login option
As long as the user needs a password for tasks like configuring WiFi /
adding a printer / installing apps, we should not allow the user
to select automatic login during FBE, as that will encourage users
to forget passwords.  Besides, during FBE, no one is likely to even
know what "automatic login" means -- better to start with a password
(and shared user) and let the user discover how to automatically login
if that is really what he/she wants.

[endlessm/eos-shell#3900]