Incorporate keyboard detection heuristic
Hooks up the dialog box ported in the previous commit to the keyboard
layout selection page. If the user clicks "Add" after the keyboard has
been guessed, it will be added to the list of keyboards on the page, and
automatically selected so that the "Next" button becomes active. If the
guessed layout was already in the list, it will still be selected.

(Roddy, Philip)

[endlessm/eos-shell#3425]