Keyboard detection trees for Spanish and Portuguese
If the user's language is Spanish or Portuguese,
assume the user is using one of the following keyboards:
br, es, latam, pt, or us.
This greatly simplifies the detection logic.
(If the user has something different, he/she can still
manually select a specific layout.)

For now, we use the same trees for Spanish and Portuguese,
under the assumption that there may be some cross-pollination
of Spanish and Portuguese speakers and keyboards throughout
Latin America.

The maps are generated via the keymaptree script provided
with the source of console-setup-1.70ubuntu10, by modifying
keymaptree.use to include only the relevant layouts.

[endlessm/eos-shell#4213]