Remove detection of us:intl keyboard
The step asking a user if the umlaut key (¨) is present
is very confusing, as it can easily be mistaken for the
double quotation marks (").  Many users who have a US keyboard
would thus incorrectly be detected as us:intl.  Besides, most
if not all users who want the us:intl layout would be using a
standard US keyboard and the umlaut would not be physically
visible on the keyboard.  Better to just let those users who
want the international mapping for the US keyboard to manually
select the appropriate layout.

[endlessm/eos-shell#4213]