language: Handle /endless overlay in test mode more robustly
The /endless overlay and SD card mount at /var/endless-extra need to be
handled specially around the other overlay mounts. Since
/var/endless-extra will be masked by the subsequent /var overlay, it
needs to be unmounted and remounted. The /endless overlay seems to
maintain connections to the real paths through the overlays, so it also
needs to be unmounted and remounted once the new paths are in place.

[endlessm/eos-shell#5036]