Add new languages
Bengali (bn): although there are upstream translations for bn_IN,
there are none for bn, even on master.

Thai (th): used existing translations, but with round-trip
to Transifex to establish a baseline.

Vietnamese (vi): although the .po file was present, it had
no translations, so I grabbed the latest translations from
upstream master and went through the round-trip to Transifex
to establish a baseline.