Introduce gnome-image-installer binary and pages
Image installer is a tool which can be used to write disk image files
on hardrives on a computer. Main usecase is to have it on a bootable
USB stick distribution and write an OS image to the hard drive of the
target computer.

No functionality yet, just initial UI pages and build structure.