Do a check for multiple partitions on target disks and warnd if there are
This is an additional hurdle to tip the user to the fact that what they
think are different drives in eg. Windows might not be. This won't prevent
accidental data erasing completely but might aid those that know they have
done a system/data split of their drive.

We check for a number of usual parition types on Windows and Linux that
hold user data, and leave protecting the more exotic ones up to the user.