Don't make gpg try to write to /usr
When you pass `--keyring X` to gpg, it seems to treat X as one of the
user's own keyrings, and tries to write a X.tmp file alongside it. This
fails because eos-image-keyring.gpg is on /usr, which is read-only; and
so image verification spuriously fails and the installer refuses to
proceed.

Other scripts use `gpgv` which does not have this behaviour, but it also
lacks progress reporting. So instead we verify the signature in two
phases:

1. Import eos-image-keyring.gpg to the gnome-initial-setup user's
   keyring, creating it if necessary.
2. Verify the image signature via the user's own keyring, additionally
   asserting that the image signing key is trusted.

I haven't added a separate step, or region of the progress bar, for the
first step, because it's essentially instantaneous.

https://phabricator.endlessm.com/T13004