convert installed system to MBR if we booted from BIOS
Run the new /usr/sbin/eos-convert-mbr script on the newly-created filesystem to
rewrite the GPT partition table as an MBR partition table, in the case that
this is a system that booted via BIOS instead of MBR. There are BIOS systems
which do not recognise the protective MBR on GPT disks as bootable, so ignore
the HDD as unbootable and others which might attempt untested EFI code paths on
hard drives, even though they might not apply the same check to our installer
USB. The most reliable and well-tested way to have these systems boot from the
HDD is use MBR partition tables on BIOS boot systems.

This change also adds a "pulsing" progress bar whilst syncfs() and this
repartitioning script is being called.

https://phabricator.endlessm.com/T12365

Modified-by: Will Thompson <wjt@endlessm.com>