polkit: grant 'live' user access if endless.live_boot-ed
polkit.spawn() is defined to throw an exception if the command returns
non-zero.

Grabbing the flag up-front and caching the result may be premature
optimization but it avoids having to contort the logic of the condition
to defer calling 'grep' until the last minute.

I don't see any security implications of adding this rule
unconditionally in all images. Even if someone creates a user called
'live' in a normal installed system, the endless.live_boot flag should
not be present. If it *is* present, then the user must have manually
edited the kernel command line, so granting a local user access to
udisks and co doesn't seem problematic.

https://phabricator.endlessm.com/T12548