diskimage: skip devices with Ignore hint set
This is set on the "real" device for the host partition (eg /dev/sda1)
which is not mountable because it is in use by device-mapper. If udisks
happens to return the real device before 'endless-image-device' (another
device-mapper volume covering the entire partition), we'd try and fail
to mount the real device and go straight to the error screen.