page: add class fields to control header behavior
The headerbar and the pages currently have a hard time
communicating with each other because we have to access
widgets directly. We can't fine tune, for example, whether
we want to show only the forward or the backward button.

This commit fine tunes the GisPage control over the header
by adding class fields that are read by GisAssistant.

https://phabricator.endlessm.com/T13754