install: hide header widgets on Install page
When the user reaches the Install page, changing to the
previous page is not possible anymore. But we keep showing
a sensitive Previous button that breaks whatever we're
doing.

The previous commit added class-wide fields to fine tune this
behavior, and this is the oportunity for us to use them.

Fix that by setting the class fields in such a way that the
back, next, and window control buttons are hidden.

https://phabricator.endlessm.com/T13754