installer: ask permission to quit reformatter while copying data
When copying data, a hard quit (e.g. using the Quit context menu
from the taskbar) means that the disk is left with possibly and
probably broken contents, and that almost certainly means unusable
computer.

Ideally, the reformatter/installer must ask if the user really wants
to do that, and explicitly say that the device might be in a broken
state after that.

Fix that by adding a confirmation dialog when the window tries to be
deleted.

https://phabricator.endlessm.com/T13754