Consistently use uint64 for installed size of images
Disk sizes are by definition strictly positive, and indeed
get_gzip_disk_image_size () and friends return unsigned values, so the
store should store an unsigned value too. This fixes a
signed-vs-unsigned comparison warning in disktarget, which compares the
uncompressed image size to the candidate target device size (which is
unsigned).

While we're here, simplify some dodgy logic dealing with the (supposedly
impossible) case where we can't determine the uncompressed size of the
image. Rather than lazily determining the uncompressed size, we store it
in the list model along with all the other facts.