Use decimal gigabytes
This matches what Nautilus and other GNOME applications show for the
size of a file or storage device, by virtue of using the same function.
This also means "GB" is localized into the interface language.

In the error case when the selected drive is too small, we show the size
of the image file in bytes, to be absolutely clear how much space is
needed.

https://phabricator.endlessm.com/T12718