standalone: install gnome-shell initial-setup mode config
Endless OS 3.1 (and earlier) included an older fork of GNOME Shell that
defined the initial-setup mode internally. The newer version of the
Shell in 3.2 removes this internal definition in favour of loading the
definition from a config file. The gnome-initial-setup package installs
this file, but the eos-installer-standalone package (which conflicts
with it) does not.

https://phabricator.endlessm.com/T17875