Omit welcome page entirely in live mode
When launched from the live FBE, it's confusing that after selecting
Reformat, there's a Back button which doesn't take you back to the
Try/Reformat page.

When launched from the live desktop, it's harmless but also useless. You
already picked a language so why have a Back button to a page that
allows you to change it?

In unattended mode, the general behaviour is that each page reads its
own settings from an ini file. Right now the language page doesn't read
this setting but it seems reasonable that in future it might be able to,
so don't omit the page in unattended mode.

Unattended && live is not currently supported.