Inhibit idle & suspend when running eos-installer
This is a "side-port" (... ?) of
https://github.com/endlessm/gnome-initial-setup/pull/188
with INHIBIT_SUSPEND thrown in for good measure. Suspending mid-way
through reformatting seems bad.

In theory, we might make this global inhibit conditional on running
within the FBE, and when within the live desktop session only inhibit
while the reformat is in progress, but this will do.

https://phabricator.endlessm.com/T18045