Improve endless.image.device regex, add assertion
I noticed the regex problem while lifting this code into
gnome-initial-setup for
<https://github.com/endlessm/gnome-initial-setup/pull/194>. There might
well be two UUIDs on the kernel command line -- one for the image
device, and one for the root fs within the image file -- so we should
take care to pick the correct one.

The g_return_val_if_fail() was suggested in code review for g-i-s, and
it's a reasonable point.