install: pass image size to gpg
When the input file is /dev/mapper/endless-image (a block device), gpg
cannot determine the file size, so (without a hint) sets the total field
of PROGRESS messages to 0.

Previously, we would substitute gis_store_get_required_size() (in bytes)
in this case. This worked with gpg1, which always gave sizes in bytes,
but gpg2's PROGRESS messages include an additional units field (KiB,
MiB, ...) and give sizes in the given unit.  (This is, according to
g10/gpg.c, to be kind to platforms with 32-bit 'int' when processing
large files.)

We ignored the units field, assumed the "cur" field was in bytes, and
divided it by the total in bytes. When "cur" is in MiB, the result is
always approximately 0, and so the progress bar does not make progress.

When given --input-size-hint, if the size cannot be determined, GPG
assumes the value given, and scales it in the same way. This, in
principle, allows gis_install_page_gpg_progress() to assume the total is
present. If it is not present, something unexpected has happened; we
handle that case, and call gtk_progress_bar_pulse() in preference to
dividing by zero.

https://phabricator.endlessm.com/T18706