disktarget: allow escape when all disks are too small
In the case where no valid disks are found, we set an error condition,
change the "Next" button to "Finish", and mark the page as complete.
This allows the user to proceed to the final page, and hence shut down
the computer.

If two disks are found, both of which are too small, then changing the
combo box from the first to the second previously caused the complete
state to be reset to FALSE, which makes the "Finish" button insensitive
and traps the user on the disktarget page.

To fix this, always use check_can_continue() to set the complete state as
needed.