disktarget: improve layout of error messages
Previously, the separator was part of the "I agree to wipe out my data"
box, not the parent. So the happy path looked like this:

    [Previous]             Select Disk              [Next]

    Select the disk you'd like to reformat with Endless OS

    ------------------------------------------------------

    Select Disk:

    [ VBOX HARDDISK ....... v ]

    ------------------------------------------------------

    [ ] I agree to erasing all of my files and apps

But the sad path looked like this, with no separator between the
dropdown and the error message:

    [Previous]             Select Disk              [Next]

    Select the disk you'd like to reformat with Endless OS

    ------------------------------------------------------

    Select Disk:

    [ VBOX HARDDISK ....... v ]

    /!\ The location you have chosen is too small ...

I think it looks better and more consistent to always show the
separator.

(My feeling is that the words "Select Disk" appear too often on this
page but that's another issue!)