Define a GisImage structure
Previously, whenever we wanted to track a new fact about images we had
to modify the diskimage C code and GtkBuilderXML, add a bunch of new
API to GisStore, and laboriously ensure the new fact is threaded all the
way through. Instead, we can encapsulate all these facts in a structure,
define a G_TYPE_BOXED for it, and just pass that around. This involves a
bit of boilerplate in gis-store.[ch] but the reward is consistent names
and types for each field throughout the application, as well as less
code everywhere else.