Verify endless.squash directly
From 3.3.0 onwards, we include .squash.asc in the ISO image and write
"<image basename>.squash" rather than "<image basename>.img" to the
'live' file. This allows us to verify the squashfs image rather than
decompressing it, verifying the decompressed contents, discarding them,
then decompressing it again to write it to disk.

As ever, this is made more complicated by the image-mapping stuff. In
fact, in the ISO case endless.squash will not be mapped. But when
creating exFAT-formatted USB sticks from an ISO, we would like to be
able to copy endless.squash, rather than decompressing it, to save space
on the USB stick. This is supported by the boot scripts, so, hooray, we
get to support it here too.

https://phabricator.endlessm.com/T17203