install: blkdiscard target device
Rather than invoking `blkdiscard` the command, which would require
using our favourite pkexec trick to escalate that process, we can just
call the two necessary ioctls.

One could argue that we should only blkdiscard the region we're about to
write, but the data after that region is spiritually overwritten, too,
with undefined contents.