install: don't assume PROGRESS message is well-formed
If the current/full values are mis-parsed, that's no huge deal (maybe
the progress bar will behave weirdly) but if there are fewer than 6
fields we would previously crash.

While we're here: log unhandled messages. It would be nice to use the
GOODSIG/BADSIG messages to report in what way the signature is invalid.