install: pass correct image size to GPG
Historically, GPG has only used the hint when we're installing from
/dev/mapper/endless-image, whose size cannot be determined with stat().
So it hasn't mattered that the "wrong" value is being used.

But when we move to reading the file and tee-ing it to GPG and the
decompressor at the same time, the distinction actually does matter,
because GPG will never be able to determine the size of its input (it's
a pipe) and will always rely on the hint.

The kink here is that when the image is a SquashFS, the compressed
size we store is the size of the SquashFS, but the data we feed to GPG
is from /dev/mapper/endless-image, which corresponds to the uncompressed
image inside the SquashFS. So there needs to be a special-case
somewhere. I chose to put it outside GisScribe because that class knows
nothing whatsoever about SquashFSes.

While we're here: g_file_info_get_size() returns a goffset, which is
signed! We store the uncompressed size as guint64, so let's store the
(always non-negative) compressed size as guint64 too. We already
implicitly assume that it's non-negative since g_format_size_full()
accepts a guint64.