install: decompress in a subprocess
On Linux, the default pipe buffer size is 64 KiB. Since we are reading
and writing 1 MiB buffers from and to each pipe, it shouldn't be too
surprising that this improves throughput. Without this detail,
decompressing in a subprocess is actually slower than just doing it in
the write thread; but with it, it's faster. (As it happens, 1 MiB is
also the maximum size an unprivileged process can set for pipe buffers.)

A nice way to explore this is to compare the runtimes of these two
commands:

                     xz -cd foo.img.xz | dd bs=1M iflag=fullblock ...
    cat foo.img.xz | xz -cd            | dd bs=1M iflag=fullblock ...

The former is what the most performant of our many shell scripts does,
and on my system with a given ~2GB base image it takes ~3:20. The latter
is closer to what this application is doing, and takes ~3:50. All pipes
here have the default buffer size. (I have inspected the cat, xz and dd
source code to confirm this.)

GisScribe with default pipe buffer sizes takes over 4 minutes (one
run took over 5 minutes, but I think this must be an outlier). With 1
MiB pipe buffers, it's around ~3:35. We can conclude that it's the pipes
into xz and GPG that are important, not the one from xz to the write
thread. This makes some sense since we are reading many more buffers
from xz than we write to it (that's the point of decompression!).

We might be able to do better by interleaving the writes in the tee
thread. Right now it's:

    while buffer = input.read(1024 * 1024):
        gpg.write_all(buffer)
        xz.write_all(buffer)

You could imagine interleaving the writes to gpg and xz, to avoid
starving xz when gpg is not ready to read:

    b = 1024 * 1024
    while buffer = input.read(b):
        x = g = b

        while x or g:
            if g:
                g -= gpg.write(buffer + b - g, g)

            if x:
                x -= xz.write(buffer  + b - x, x)

But let's stick with "straightforward and reasonably performant".