Rename desktop file to com.endlessm.Installer
This matches the GApplication application ID (and icon name). This
change is necessary to inhibit shutdown while the reformatter is
running: when showing inhibitors in the shutdown dialog, GNOME Shell has
only the GApplication-style application ID to go by. It attempts to find
the corresponding application and, if it fails, just hides the
inhibitor.

https://phabricator.endlessm.com/T20103