finished: remove support email address from translated string
For some languages (es, pt, id) we use a different customer support
email address; for all other languages we use support@endlessm.com.
Previously we handled this by including the email address in the string
to be translated, changing the address in those three translations, and
relying on translators for other languages to leave it unmodified.

However, Transifex now does not allow the target of the link to be
edited in the translation. The string would show up in Transifex as:

  Please contact <1>support@endlessm.com<1> or join the <2>Endless
  Community<2> to troubleshoot.

where <1> and <2> are special symbols that the Transifex user cannot
manipulate. For the forum link this is desirable, but for the mailto:
link it is not.

In the Shell, we already have a different mechanism to vary the support
email address by locale. So, we can reuse this here, and replace the
email link in the translatable string with a placeholder.

https://phabricator.endlessm.com/T19689