unattended: add confirmation page
This layout is modelled on the way GNOME Disks shows disk properties and
Settings → About… shows properties of the computer. We show the disk
vendor & model as one field for two reasons:

1. This matches the dropdown in the 'disktarget' page
2. On every machine I have tried, the 'Vendor' property of the
   UDisks2.Drive object is blank, and all information is in the 'Model'
   property

The fields showing the current vendor and product are placeholders; they
will be implemented in a future commit.

https://phabricator.endlessm.com/T14343