unattended: introduce GisUnattendedConfig stub
At present, the keyfile is parsed up-front into a GKeyFile, which is
stashed in the store, and each page plucks keys out of it. This works,
but it makes it harder to see what information is retrieved from the
file (since it's spread out through the program) and means the keyfile
cannot be validated up-front.

We will gradually move this into this class, improving and extending the
file format as we go.

https://phabricator.endlessm.com/T14343