unattended: allow matching 0, 1, or more computers
Previously, exactly one computer had to be specified in the config file;
if the current computer did not match, the installer would silently
revert to attended mode. The new implementation proceeds in unattended
mode in that case, but presents a warning to the operator and requires a
(single-click) confirmation before reformatting begins.

The new implementation also allows zero computers to be specified; in
this case, no warning is shown (this is not an error) but confirmation
is still required.

We also change the DMI data sanitation a little to strip leading and
trailing whitespace, reject strings which are empty after sanitation,
and preserve case. The latter is largely an aesthetic choice: "innotek
GmbH", "VirtualBox", "Dell Inc." and "XPS 13 9343" look better than the
lower-cased versions in a dialog with many titlecased strings. The
matching logic is still case-insensitive.

https://phabricator.endlessm.com/T14343