diskimage: set GIS_STORE_IMAGE_DRIVE sooner
This value is used by the disktarget page (which follows diskimage) to
avoid offering to overwrite the disk the image file is stored on. If the
image partition is already mounted, gis_diskimage_page_populate_model()
was previously called before setting GIS_STORE_IMAGE_DRIVE, and in
unattended mode this function advances to the next page *before*
returning. This meant that GIS_STORE_IMAGE_DRIVE was previously not set
at the point it was accessed.

(In practice this would never happen because the .mount unit that is
supposed to mount the eosimages partition does not work, which we'll fix
shortly.)

https://phabricator.endlessm.com/T14343