unattended: support writing config
If you, dear reader, are sad about functions with loads of string
parameters, so am I. The alternative would be to make it possible to
create an empty GisUnattendedConfig, fill in fields with named setters,
then save it. But currently GisUnattendedConfig is immutable once it has
been intialized, and I want to preserve that.

https://phabricator.endlessm.com/T14343