install: don't unmount image partition
We want to be able to write things here from the 'finished' page, namely
a new unattended.ini (or perhaps a log file on error, in future). The
partition will be cleanly unmounted when the system is shut down, and in
any case there's minimal risk in an unclean unmount.