Allow going Back from first page when run from FBE
This allows the user to get back to the Try/Reformat choice page (in the
combined USB/ISO case) or the Language page (in the eosinstaller image
case).

It's enough to just make the process exit cleanly, since in FBE mode
there is no other way this can happen: on the final page, there's just a
Power Off button, and this will (indirectly) cause the process to be
killed. The FBE has been modified to take the correct action when this
process exits cleanly.

https://phabricator.endlessm.com/T14616