disktarget: don't manipulate list of pages
Previously, this removed the 'install' page, which happens to be the
penultimate page, from a GList borrowed from GisAssistant. I guess this
is because we used to show a row of •s at the bottom of the window to
indicate the number of pages and the current position within them, but
we don't show that any more. The 'install' page skips itself if
gis_store_get_error() returns non-NULL.

https://phabricator.endlessm.com/T19950