tests: add missing #include for memfd_create()
glibc >= 2.27 define this syscall. 025d89e updated our libglnx snapshot
to a version which includes this header to check for the existing
definition before defining a new version; but didn't update the function
which uses memfd_create() to include this header.

(The header exists in glibc < 2.27, so it's harmless to include it
there.)