Adapt to GCC 8's -Wcast-function-type
This new diagnostic produces warnings like:

  gis-scribe.c:1075:25: error: cast between incompatible function types
    from ‘gboolean (*)(GObject *, void *)’ {aka ‘int (*)(struct _GObject *, void *)’}
    to ‘gboolean (*)(void *)’ {aka ‘int (*)(void *)’} [-Werror=cast-function-type]

g_task_attach_source() takes a GSourceFunc which it passes to
g_source_set_callback(). However, the actual function type depends on
the specific GSource's implementation; the function pointer will be cast
back to the correct type later. So in this case the warning is (a)
spurious; and (b) forced on us by the GLib API. The workaround described
in the GCC documentation is to cast via (void (*) (void)); we employ
that here.

https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html#index-Wcast-function-type

For the cast to GtkTickCallback, the warning is legit: the function did
have the wrong type. It was leaning on my belief that on all platforms
we care about, the ABI allows calling a function with extra arguments.
But it's easy to declare the function with the right type, and remove
the cast.