main: avoid NULL-pointer dereference
Recent GCC warns about this.

(In fact, this function is an unused vestige of gnome-initial-setup. The
language will never be changed, so pages will only be "rebuilt" once, so
there will never be a current page, so destroy_pages_after() will never
be called. But this one-line fix will do until we have time to unpick
this unused code.)