scribe: check expected number of bytes are teed
I saw a transient GPG verification failure when reformatting from a
nightly ISO in a virtual machine. There was no useful information in the
journal (beyond GPG's "BAD signature" output); re-running the process
(without a reboot or anything) worked fine; and I manually GPG-verified
the ISO, the squashfs within it, and the uncompressed image within that.
Mysterious!

In the absence of any clues, here's one coherence check we could add. In
theory this is redundant with the similar check on the total
uncompressed size in the write thread -- particularly on an ISO, where
we're verifying and writing the loop-mounted uncompressed image so the
two sizes are equal. But maintaining two additional counters shouldn't
hurt too much.