scribe: ensure first subtask error is reported
I had a report of an installation failing with:

> error writing image to self: Error writing to file descriptor: Broken
> pipe

which comes from the tee thread. I believe what actually happened is the
writer thread failed because the target hard disk is failing (as
indicated by a pop-up notification and the SMART information in GNOME
Disks), and we just picked the wrong error to display. Unfortunately the
message about subtask errors we're discarding was g_debug() and so would
not have appeared in the journal, even if I had been able to get the
journal from the reporter.

Looking at the code, on write error gis_scribe_write_thread() will call
g_task_return_error(), then close() the read end of the pipe coming from
the decompressor/tee thread, set its task as failed, and return. The tee
thread, on error, does something similar.

But when a task is running in a thread, g_task_return_error() doesn't
actually schedule the callback until the thread function returns. I
think this means that if the write thread happens to yield between
calling close() on the pipe from the tee thread, it's possible that the
tee task will signal its error on the main thread first, and we'd report
a less useful error.

To fix this, make each subtask explicitly set the error when it fails.
We also log all errors reported by subtasks when the main thread hears
about them.

All access to self->error was already guarded by the glorious mutex
because the write thread needs to wait for all other threads to complete
with success.