scribe: Allow verifying image with sha256 checksum
As an alternative to requiring a GPG signature from a trusted key, allow
verifying the image with a sha256 checksum. The checksum file is
expected to reside alongside the image with a `.sha256` suffix.
Typically this would be a file formatted by the `sha256sum` program, but
any file where the first word is a sha256 hex digest is acceptable.

While this seems less secure, in reality it provides no more validation
than using GPG since the trusted key is embedded in the payload. An
attacker could alter the signature and keyring to make it appear that it
came from the expected party. So, all that the signature or checksum can
truly validate is that the image file contains the bytes the person
providing it expected.

https://phabricator.endlessm.com/T31304