scribe: Fix deadlock from async initialization errors
If there is an error setting up any of the tasks in
`gis_scribe_write_async`, it needs to be propagated to
`GisScribe::error`. This requires locking the mutex, but currently
that's already locked throughout `gis_scribe_write_async` to guard
changes to `GisScribe::outstanding_tasks`. This causes a deadlock if the
various `gis_scribe_begin_*` setup functions try to set an error since
they won't return.

A concrete example of this is if an image signature file doesn't exist.
The installer will instead show no progress until it's killed. Instead,
lock and unlock the mutex around each access to
`GisScribe::outstanding_tasks`. Another option would be to use a
separate locking mechanism for the outstanding tasks and error, but this
seemed simpler.

https://phabricator.endlessm.com/T31304