disktarget: Skip read-only block devices
If the block device for the prospective drive or loop target is
read-only, don't offer it as an installation target. This is
particularly helpful in the ISO or live USB case where the source media
is mounted from a read-only loop device and would otherwise be presented
as a possible installation target.

https://phabricator.endlessm.com/T31304