disktarget: Skip loops with missing backing files
If a loop device's backing file doesn't exist, then assume that the loop
can't be used as an installation target. This is helpful on the live
media where the loop devices are mounted in the initramfs, so checking
if they're mounted in the current mount namespace wouldn't help.
However, since the backing files also aren't in the current mount
namespace, then we can assume that they're used somewhere else.

https://phabricator.endlessm.com/T31304